HyperplanesAndMatroids

Documentation for HyperplanesAndMatroids.

HyperplanesAndMatroids.SignVector

HyperplanesAndMatroids.SignVector

Base.show

HyperplanesAndMatroids.negativePart

HyperplanesAndMatroids.positivePart

HyperplanesAndMatroids.support

HyperplanesAndMatroids.zeroPart

HyperplanesAndMatroids.SignVector — Type

SignVector

A SignVector represents a vector w/ real entries, designed to be equal if and only if they have the same signs

Example:

julia> SignVector([0, 1, -1, -1]) 0+--

HyperplanesAndMatroids.SignVector — Method

SignVector(str::String)

Construct a sign vector from a string with characters '0', '+', '-'

Examples:

julia> SignVector("0+--") 0+--

Base.show — Method

Base.show(io::IO, X::SignVector)

Pretty printing for sign vectors

HyperplanesAndMatroids.negativePart — Method

negativePart(X::SignVector)

Returns the positive part of a sign vector, i.e. {i ∣ X_i = -}

HyperplanesAndMatroids.positivePart — Method

positivePart(X::SignVector)

Returns the positive part of a sign vector, i.e. {i ∣ X_i = +}

HyperplanesAndMatroids.support — Method

support(X::SignVector)

Returns the support of a sign vector, i.e. {i ∣ X_i ≠ 0}

HyperplanesAndMatroids.zeroPart — Method

zeroPart(X::SignVector)

Returns the zero part of a sign vector, i.e. {i ∣ X_i = 0}